+++ mobile/android/android-components/docs/changelog.md
@@ -35,6 +35,9 @@ permalink: /changelog/
 * **service-nimbus**
   * Added `getExperimentBranches` method to `Nimbus` for retrieving a list of experiment branches for a given experiment. [issue #9895](https://github.com/mozilla-mobile/android-components/issues/9895)
 
+* **feature-tabs**
+  * Added usecase for duplicating tabs to `TabsUseCases`.
+
 # 74.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v73.0.0...v74.0.0)